Further updated the network self-healing code. Still needs some work. Nodes are successfully negotiating enough to allow a node to takeover, and when a node gets the votes of all parties involved it coalesces its DHT area with the dead node. This coalescing currently only works when the two DHT areas can be made into one big square. This is because to handle the case where they don't fit so nicely together, a node will have to have the ability to handle an arbitrary number of disjoint DHT areas at a time and that requires a lot of rework of a lot of code. So I'm dealing with this case later. But even beyond that, there's a couple things that need to be dealt with:
o For some reason a vector erase call is made to an invalid iterator sometime after the dead node is coalesced in. I believe this call is being made by the node that took over the dead node's area. I've looked at every erase call being made and I can't imagine which one it could be. But I've thought of outputting a debug statement before each one is made to check.

o When a node gets the votes of all the involved parties to take over a node, it notifies all those parties that it is about to do so. But the code isn't yet written to handle this notification.

o I need to think more about the situation in which a node has sent out a request to takeover a node and then gets a request to takeover the same node. This comes up often. Really I need to carefully consider and think over the entire negotiation algorithm, because there's a lot of special cases and situations to consider.

o Finally I just want to make sure that when a node takes over a dead node's area that this is properly displayed in the DHT area visualizer.